pyproject.toml
[tool.poetry]
name = "synapse-selector"
version = "0.1.16b43"
description = "Small internal package for post-selection of Glutmate/Calcium signal responses."
authors = ["Stephan Weissbach <s.weissbach@uni-mainz.de>"]
readme = "README.md"
packages = [{include = "synapse_selector"}]

[tool.poetry.dependencies]
python = ">=3.9,<3.13"
numpy = "^1.26.1"
pandas = "^2.1.1"
scipy = "^1.11.3"
plotly = "^5.17.0"
openpyxl = "^3.1.2"
torch = "^2.1.1"
pyqt6-webengine = "^6.6.0"
requests = "^2.31.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
synapse_selector = 'synapse_selector.main:main'
